LPS-159013 Revamp PortletPathsUtil into PortletRenderUtil
This is to be able to type the Map and also to give a more meaningful name to
what the class does.
