+++ portal-impl/src/com/liferay/portal/layoutconfiguration/util/PortletRenderer.java
@@ -19,7 +19,8 @@ import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Portlet;
 import com.liferay.portal.kernel.portlet.PortletContainerException;
 import com.liferay.portal.kernel.portlet.PortletContainerUtil;
-import com.liferay.portal.kernel.portlet.PortletPathsUtil;
+import com.liferay.portal.kernel.portlet.render.PortletRenderParts;
+import com.liferay.portal.kernel.portlet.render.PortletRenderUtil;
 import com.liferay.portal.kernel.servlet.BufferCacheServletResponse;
 import com.liferay.portal.kernel.util.WebKeys;
 
@@ -149,25 +150,25 @@ public class PortletRenderer {
 			new BufferCacheServletResponse(httpServletResponse);
 
 		try {
-			Map<String, Object> paths = null;
+			PortletRenderParts portletRenderParts = null;
 
 			if (_columnId == null) {
 				httpServletRequest.setAttribute(
 					WebKeys.RENDER_PORTLET_RESOURCE, Boolean.TRUE);
 
-				paths = PortletPathsUtil.getPortletPaths(
+				portletRenderParts = PortletRenderUtil.getPortletRenderParts(
 					httpServletRequest, StringPool.BLANK, _portlet);
 
-				PortletPathsUtil.writeHeaderPaths(
-					bufferCacheServletResponse, paths);
+				PortletRenderUtil.writeHeaderPaths(
+					bufferCacheServletResponse, portletRenderParts);
 			}
 
 			PortletContainerUtil.render(
 				httpServletRequest, bufferCacheServletResponse, _portlet);
 
-			if (paths != null) {
-				PortletPathsUtil.writeFooterPaths(
-					bufferCacheServletResponse, paths);
+			if (portletRenderParts != null) {
+				PortletRenderUtil.writeFooterPaths(
+					bufferCacheServletResponse, portletRenderParts);
 			}
 
 			return bufferCacheServletResponse.getStringBundler();
