+++ portal-kernel/src/com/liferay/portal/kernel/portlet/PortletPathsUtil.java
@@ -15,25 +15,17 @@
 package com.liferay.portal.kernel.portlet;
 
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.model.LayoutTypePortlet;
 import com.liferay.portal.kernel.model.Portlet;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.portlet.render.PortletRenderParts;
+import com.liferay.portal.kernel.portlet.render.PortletRenderUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
-import com.liferay.portal.kernel.util.HtmlUtil;
-import com.liferay.portal.kernel.util.HttpComponentsUtil;
-import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.WebKeys;
 
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.io.Writer;
 
-import java.util.LinkedHashSet;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import javax.portlet.MimeResponse;
 
@@ -42,253 +34,33 @@ import javax.servlet.http.HttpServletResponse;
 
 /**
  * @author Raymond Augé
+ * @deprecated As of Cavanaugh (7.4.x), replaced with {@link PortletRenderUtil}
  */
+@Deprecated
 public class PortletPathsUtil {
 
+	/**
+	 * @deprecated As of Cavanaugh (7.4.x), replaced with {@link PortletRenderUtil#getPortletRenderParts(HttpServletRequest, String, Portlet)}
+	 */
+	@Deprecated
 	public static Map<String, Object> getPortletPaths(
 		HttpServletRequest httpServletRequest, String portletHTML,
 		Portlet portlet) {
 
-		boolean portletOnLayout = false;
-
-		if (portlet.isInstanceable()) {
-			String rootPortletId = _getRootPortletId(portlet);
-			String portletId = portlet.getPortletId();
-
-			for (Portlet layoutPortlet : _getAllPortlets(httpServletRequest)) {
-
-				// Check to see if an instance of this portlet is already in the
-				// layout, but ignore the portlet that was just added
-
-				String layoutPortletRootPortletId = _getRootPortletId(
-					layoutPortlet);
-
-				if (rootPortletId.equals(layoutPortletRootPortletId) &&
-					!portletId.equals(layoutPortlet.getPortletId())) {
-
-					portletOnLayout = true;
-
-					break;
-				}
-			}
-		}
-
-		return _getPortletPaths(
-			httpServletRequest, portletHTML, portlet, portletOnLayout);
-	}
-
-	public static void writeFooterPaths(
-			HttpServletResponse httpServletResponse, Map<String, Object> paths)
-		throws IOException {
-
-		_writePaths(
-			httpServletResponse, (Set<String>)paths.get("footerCssPaths"),
-			(Set<String>)paths.get("footerJavaScriptPaths"));
-	}
-
-	public static void writeHeaderPaths(
-			HttpServletResponse httpServletResponse, Map<String, Object> paths)
-		throws IOException {
-
-		_writePaths(
-			httpServletResponse, (Set<String>)paths.get("headerCssPaths"),
-			(Set<String>)paths.get("headerJavaScriptPaths"));
-	}
-
-	public static void writeJavaScriptPath(
-			Writer writer, String javaScriptPath,
-			Map<String, String> attributes)
-		throws IOException {
-
-		String type = "text/javascript";
-
-		if (javaScriptPath.startsWith("module:")) {
-			javaScriptPath = javaScriptPath.substring(7);
-
-			type = "module";
-		}
-
-		PrintWriter printWriter = new PrintWriter(writer, true);
-
-		printWriter.print("<script src=\"");
-		printWriter.print(HtmlUtil.escapeAttribute(javaScriptPath));
-		printWriter.print("\" type=\"");
-		printWriter.print(type);
-		printWriter.print(StringPool.QUOTE);
-
-		if (attributes != null) {
-			for (Map.Entry<String, String> entry : attributes.entrySet()) {
-				printWriter.print(StringPool.SPACE);
-				printWriter.print(entry.getKey());
-				printWriter.print("=\"");
-				printWriter.print(HtmlUtil.escapeAttribute(entry.getValue()));
-				printWriter.print(StringPool.QUOTE);
-			}
-		}
-
-		printWriter.println("></script>");
-	}
-
-	private static List<Portlet> _getAllPortlets(
-		HttpServletRequest httpServletRequest) {
-
-		List<Portlet> allPortlets =
-			(List<Portlet>)httpServletRequest.getAttribute(
-				WebKeys.ALL_PORTLETS);
-
-		if (ListUtil.isNotEmpty(allPortlets)) {
-			return allPortlets;
-		}
-
-		ThemeDisplay themeDisplay =
-			(ThemeDisplay)httpServletRequest.getAttribute(
-				WebKeys.THEME_DISPLAY);
-
-		LayoutTypePortlet layoutTypePortlet =
-			themeDisplay.getLayoutTypePortlet();
-
-		allPortlets = layoutTypePortlet.getAllPortlets();
-
-		httpServletRequest.setAttribute(WebKeys.ALL_PORTLETS, allPortlets);
-
-		return allPortlets;
-	}
-
-	private static Map<String, Object> _getPortletPaths(
-		HttpServletRequest httpServletRequest, String portletHTML,
-		Portlet portlet, boolean portletOnLayout) {
-
-		Set<String> footerCssSet = new LinkedHashSet<>();
-		Set<String> footerJavaScriptSet = new LinkedHashSet<>();
-		Set<String> headerCssSet = new LinkedHashSet<>();
-		Set<String> headerJavaScriptSet = new LinkedHashSet<>();
-
-		if (!portletOnLayout && portlet.isAjaxable()) {
-			Portlet rootPortlet = portlet.getRootPortlet();
-
-			for (String footerPortalCss : portlet.getFooterPortalCss()) {
-				if (!HttpComponentsUtil.hasProtocol(footerPortalCss)) {
-					footerPortalCss =
-						PortalUtil.getPathContext() + footerPortalCss;
-
-					footerPortalCss = PortalUtil.getStaticResourceURL(
-						httpServletRequest, footerPortalCss,
-						rootPortlet.getTimestamp());
-				}
-
-				footerCssSet.add(footerPortalCss);
-			}
-
-			for (String footerPortalJavaScript :
-					portlet.getFooterPortalJavaScript()) {
-
-				if (!HttpComponentsUtil.hasProtocol(footerPortalJavaScript)) {
-					footerPortalJavaScript =
-						PortalUtil.getPathContext() + footerPortalJavaScript;
-
-					footerPortalJavaScript = PortalUtil.getStaticResourceURL(
-						httpServletRequest, footerPortalJavaScript,
-						rootPortlet.getTimestamp());
-				}
-
-				footerJavaScriptSet.add(footerPortalJavaScript);
-			}
-
-			for (String footerPortletCss : portlet.getFooterPortletCss()) {
-				if (!HttpComponentsUtil.hasProtocol(footerPortletCss)) {
-					footerPortletCss =
-						portlet.getStaticResourcePath() + footerPortletCss;
-
-					footerPortletCss = PortalUtil.getStaticResourceURL(
-						httpServletRequest, footerPortletCss,
-						rootPortlet.getTimestamp());
-				}
-
-				footerCssSet.add(footerPortletCss);
-			}
-
-			for (String footerPortletJavaScript :
-					portlet.getFooterPortletJavaScript()) {
-
-				if (!HttpComponentsUtil.hasProtocol(footerPortletJavaScript)) {
-					footerPortletJavaScript =
-						portlet.getStaticResourcePath() +
-							footerPortletJavaScript;
-
-					footerPortletJavaScript = PortalUtil.getStaticResourceURL(
-						httpServletRequest, footerPortletJavaScript,
-						rootPortlet.getTimestamp());
-				}
-
-				footerJavaScriptSet.add(footerPortletJavaScript);
-			}
-
-			for (String headerPortalCss : portlet.getHeaderPortalCss()) {
-				if (!HttpComponentsUtil.hasProtocol(headerPortalCss)) {
-					headerPortalCss =
-						PortalUtil.getPathContext() + headerPortalCss;
-
-					headerPortalCss = PortalUtil.getStaticResourceURL(
-						httpServletRequest, headerPortalCss,
-						rootPortlet.getTimestamp());
-				}
-
-				headerCssSet.add(headerPortalCss);
-			}
-
-			for (String headerPortalJavaScript :
-					portlet.getHeaderPortalJavaScript()) {
-
-				if (!HttpComponentsUtil.hasProtocol(headerPortalJavaScript)) {
-					headerPortalJavaScript =
-						PortalUtil.getPathContext() + headerPortalJavaScript;
-
-					headerPortalJavaScript = PortalUtil.getStaticResourceURL(
-						httpServletRequest, headerPortalJavaScript,
-						rootPortlet.getTimestamp());
-				}
-
-				headerJavaScriptSet.add(headerPortalJavaScript);
-			}
-
-			for (String headerPortletCss : portlet.getHeaderPortletCss()) {
-				if (!HttpComponentsUtil.hasProtocol(headerPortletCss)) {
-					headerPortletCss =
-						portlet.getStaticResourcePath() + headerPortletCss;
-
-					headerPortletCss = PortalUtil.getStaticResourceURL(
-						httpServletRequest, headerPortletCss,
-						rootPortlet.getTimestamp());
-				}
-
-				headerCssSet.add(headerPortletCss);
-			}
-
-			for (String headerPortletJavaScript :
-					portlet.getHeaderPortletJavaScript()) {
-
-				if (!HttpComponentsUtil.hasProtocol(headerPortletJavaScript)) {
-					headerPortletJavaScript =
-						portlet.getStaticResourcePath() +
-							headerPortletJavaScript;
-
-					headerPortletJavaScript = PortalUtil.getStaticResourceURL(
-						httpServletRequest, headerPortletJavaScript,
-						rootPortlet.getTimestamp());
-				}
-
-				headerJavaScriptSet.add(headerPortletJavaScript);
-			}
-		}
+		PortletRenderParts portletRenderParts =
+			PortletRenderUtil.getPortletRenderParts(
+				httpServletRequest, portletHTML, portlet);
 
 		return HashMapBuilder.<String, Object>put(
-			"footerCssPaths", footerCssSet
+			"footerCssPaths", portletRenderParts.getFooterCssPaths()
 		).put(
-			"footerJavaScriptPaths", footerJavaScriptSet
+			"footerJavaScriptPaths",
+			portletRenderParts.getFooterJavaScriptPaths()
 		).put(
-			"headerCssPaths", headerCssSet
+			"headerCssPaths", portletRenderParts.getHeaderCssPaths()
 		).put(
-			"headerJavaScriptPaths", headerJavaScriptSet
+			"headerJavaScriptPaths",
+			portletRenderParts.getHeaderJavaScriptPaths()
 		).put(
 			"markupHeadElements",
 			() -> {
@@ -304,44 +76,48 @@ public class PortletPathsUtil {
 				return null;
 			}
 		).put(
-			"portletHTML", portletHTML
+			"portletHTML", portletRenderParts.getPortletHTML()
 		).put(
-			"refresh", !portlet.isAjaxable()
+			"refresh", portletRenderParts.isRefresh()
 		).build();
 	}
 
-	private static String _getRootPortletId(Portlet portlet) {
-
-		// Workaround for portlet#getRootPortletId because that does not return
-		// the proper root portlet ID for OpenSocial and WSRP portlets
-
-		Portlet rootPortlet = portlet.getRootPortlet();
-
-		return rootPortlet.getPortletId();
-	}
-
-	private static void _writePaths(
-			HttpServletResponse httpServletResponse, Set<String> cssPathsSet,
-			Set<String> javaScriptPathsSet)
+	/**
+	 * @deprecated As of Cavanaugh (7.4.x), replaced with {@link PortletRenderUtil#writeFooterPaths(HttpServletResponse, PortletRenderParts)}
+	 */
+	@Deprecated
+	public static void writeFooterPaths(
+			HttpServletResponse httpServletResponse, Map<String, Object> paths)
 		throws IOException {
 
-		if ((cssPathsSet == null) || (javaScriptPathsSet == null) ||
-			(cssPathsSet.isEmpty() && javaScriptPathsSet.isEmpty())) {
+		PortletRenderUtil.writeFooterPaths(
+			httpServletResponse,
+			new PortletRenderParts(
+				(Collection<String>)paths.get("footerCssPaths"),
+				(Collection<String>)paths.get("footerJavaScriptPaths"),
+				(Collection<String>)paths.get("headerCssPaths"),
+				(Collection<String>)paths.get("headerJavaScriptPaths"),
+				(String)paths.get("portletHTML"),
+				(Boolean)paths.get("refresh")));
+	}
 
-			return;
-		}
+	/**
+	 * @deprecated As of Cavanaugh (7.4.x), replaced with {@link PortletRenderUtil#writeHeaderPaths(HttpServletResponse, PortletRenderParts)}
+	 */
+	@Deprecated
+	public static void writeHeaderPaths(
+			HttpServletResponse httpServletResponse, Map<String, Object> paths)
+		throws IOException {
 
-		PrintWriter printWriter = httpServletResponse.getWriter();
-
-		for (String cssPath : cssPathsSet) {
-			printWriter.print("<link href=\"");
-			printWriter.print(HtmlUtil.escape(cssPath));
-			printWriter.println("\" rel=\"stylesheet\" type=\"text/css\" />");
-		}
-
-		for (String javaScriptPath : javaScriptPathsSet) {
-			writeJavaScriptPath(printWriter, javaScriptPath, null);
-		}
+		PortletRenderUtil.writeHeaderPaths(
+			httpServletResponse,
+			new PortletRenderParts(
+				(Collection<String>)paths.get("footerCssPaths"),
+				(Collection<String>)paths.get("footerJavaScriptPaths"),
+				(Collection<String>)paths.get("headerCssPaths"),
+				(Collection<String>)paths.get("headerJavaScriptPaths"),
+				(String)paths.get("portletHTML"),
+				(Boolean)paths.get("refresh")));
 	}
 
 }
