+++ portal-kernel/src/com/liferay/portal/kernel/portlet/render/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
