+++ portal-web/docroot/html/common/themes/init.jsp
@@ -18,7 +18,7 @@
 
 <%@ page import="com.liferay.portal.kernel.model.Portlet" %><%@
 page import="com.liferay.portal.kernel.model.portlet.PortletDependency" %><%@
-page import="com.liferay.portal.kernel.portlet.PortletPathsUtil" %><%@
+page import="com.liferay.portal.kernel.portlet.render.PortletRenderUtil" %><%@
 page import="com.liferay.portal.kernel.servlet.BrowserMetadata" %><%@
 page import="com.liferay.portal.kernel.upload.UploadServletRequestConfigurationHelperUtil" %><%@
 page import="com.liferay.portal.servlet.ComboServletStaticURLGenerator" %><%@
