+++ portal-web/docroot/html/common/themes/top_portlet_resources_js.jspf
@@ -30,7 +30,7 @@
 
 	<%
 	for (String url : urls) {
-		PortletPathsUtil.writeJavaScriptPath(out, url, Collections.singletonMap("data-senna-track", "temporary"));
+		PortletRenderUtil.writeJavaScriptPath(out, url, Collections.singletonMap("data-senna-track", "temporary"));
 	}
 	%>
